Changed how the expanding bar is managed
Previous method used a single instance of polybar moved up and down.
This didn't work because the bar was either always above every window
(which breaks fullscreen mode) or always below every window (so it is
obscured when expanded). New method uses to separate instances of
polybar (one expanded and one collapsed) that are shown and hidden as
appropriate.