Rofi: Readded icons into window switch script
Added icons back into the window switch script in a (slightly) more
elegant manner that doesn't involve searching .desktop files. It uses
only the .app_id field (or .window_properties.instance for XWayland) to
select the icon. This method works on the assumption that the icon name
closely matches the app_id which is true for most applications but not
all. E.g. pavucontrol's app_id is pavucontrol but the icon should be
multimedia-volume-control so this method fails.